Keep all FileDetectors in file_detector.py module.
